dragging: add support for multiple segment drag
If a touch event happens shortly after a drag-and-drop event, it is
handled specially. Specifically, the movement since the last
drag-end-drop has ended is sent as a single jump, then a new
drag-and-drop starts.

The real use case for that is drawing a single thing with multiple
strokes, i.e. writing a formula

By default the threshold is 0ms, so the feature is off